создание объекта IOManager и его интеграция в други комманды

+++ src/main/kotlin/WorkerClass/Coordinates.kt
@@ -1,7 +1,7 @@
 package WorkerClass
 
 class Coordinates(
-    val x: Long, val y: Double
+    val x: Float, val y: Long
 ) {
     override fun toString(): String {
         return "Coordinates(x=$x, y=$y)"
